simulation of a pulse in free space in the transverse magnetic mode

+++ fd2d/clang/test_3_1.c
@@ -11,7 +11,7 @@
 
 
 float gaussian(int t, int t0, float sigma) {
-    return exp(-0.5*(t - t0)/sigma*(t - t0)/sigma);
+    return expf(-0.5f*(t - t0)/sigma*(t - t0)/sigma);
 }
 
 
@@ -21,7 +21,7 @@ void dfield(int t, int nx, int ny, float *dz, float *hx, float *hy) {
     for (int i = 1; i < nx; i++) {
         for (int j = 1; j < ny; j++) {
             int n = i*ny+j;
-            dz[n] += 0.5 * (hy[n] - hy[n-ny] - hx[n] + hx[n-1]);
+            dz[n] += 0.5f * (hy[n] - hy[n-ny] - hx[n] + hx[n-1]);
         }
     }
     /* put a Gaussian pulse in the middle */
@@ -47,8 +47,8 @@ void hfield(int nx, int ny, float *ez, float *hx, float *hy) {
     for (int i = 0; i < nx-1; i++) {
         for (int j = 0; j < ny-1; j++) {
             int n = i*ny+j;
-            hx[n] += 0.5 * (ez[n] - ez[n+1]);
-            hy[n] -= 0.5 * (ez[n] - ez[n+ny]);
+            hx[n] += 0.5f * (ez[n] - ez[n+1]);
+            hy[n] -= 0.5f * (ez[n] - ez[n+ny]);
         }
     }
 }
